# api/swagger-schemas.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Your API",
    "description": "API Documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "title": { "type": "string" },
          "price": { "type": "number" },
          "description": { "type": "string" },
          "category": { "type": "string" },
          "image": { "type": "string" },
          "rating": {
            "type": "object",
            "properties": {
              "rate": { "type": "number" },
              "count": { "type": "number" }
            },
            "required": ["rate", "count"]
          }
        },
        "required": [
          "id",
          "title",
          "price",
          "description",
          "category",
          "image",
          "rating"
        ]
      },
      "CartItem": {
        "type": "object",
        "properties": {
          "productId": { "type": "number" },
          "userId": { "type": "string" }
        },
        "required": ["productId", "userId"]
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/products": {
      "get": {
        "summary": "Get all products",
        "tags": ["Products"],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Product" }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/addToCart/{userId}/{productId}": {
      "post": {
        "summary": "Add product to cart",
        "tags": ["Cart"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the user"
          },
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "The ID of the product to add to the cart"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Product added to cart successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/cartItems/{userId}": {
      "get": {
        "summary": "Get all cart items for a user",
        "tags": ["Cart"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the user to get cart items for"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved cart items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CartItem" }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/deleteFromCart/{userId}/{productId}": {
      "delete": {
        "summary": "Delete product from cart",
        "tags": ["Cart"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the user"
          },
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "The ID of the product to remove from the cart"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Product removed from cart successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Product or Cart not found"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}